+++ src/pages/password/ui/reset-password-page/index.tsx
@@ -7,9 +7,15 @@ import { toast } from 'react-toastify'
 
 import { QueryErrorBoundary } from 'shared/api'
 import { isEmptyString } from 'shared/lib'
-import { Button, HelperText, Label, Layout } from 'shared/ui'
+import { Button, Label, Layout } from 'shared/ui'
 
-import { field, fieldWrapper, Form, helperText } from 'features/form'
+import {
+  confirmPasswordRules,
+  passwordRules,
+  PasswordValidation,
+} from 'entities/auth'
+
+import { field, fieldWrapper, Form } from 'features/form'
 
 import { useCheckCodeValidity } from '../../api/use-check-code-validity'
 import { useResetPassword } from '../../api/use-reset-password'
@@ -17,13 +23,6 @@ import {
   resetFormDefaultValues,
   ResetFormValues,
 } from '../../model/form-values'
-import {
-  hasLowercaseAndUppercaseLetter,
-  hasNumber,
-  hasSpecialChar,
-  isCorrectLength,
-} from '../../model/rules'
-import * as rules from '../../model/rules'
 import { ErrorFallback } from '../error-fallback'
 
 import * as css from './variants'
@@ -49,16 +48,9 @@ const ResetPasswordPage = () => {
 
   const canSubmit = !isEmptyString(password) && !isEmptyString(confirmPassword)
 
-  const checkPasswordCondition = (condition: (value: string) => boolean) => {
-    if (isEmptyString(password)) return 'default'
-    if (condition(password)) return 'success'
-
-    return 'error'
-  }
-
   const handleResetPasswordSuccess = () => {
     toast.success('Your password has been changed successfully')
-    router.push('/sign-in')
+    router.replace('/sign-in')
   }
 
   const submitForm = ({ password }: ResetFormValues) => {
@@ -84,42 +76,18 @@ const ResetPasswordPage = () => {
             <div className={field()}>
               <Form.PasswordField
                 name="password"
-                rules={rules.password}
+                rules={passwordRules}
                 placeholder="Enter the password"
                 autoComplete="one-time-code"
                 showToggle
               />
-              <div className={helperText()}>
-                <HelperText
-                  hasIcon
-                  variant={checkPasswordCondition(isCorrectLength)}
-                >
-                  9 ~ 28 characters long
-                </HelperText>
-                <HelperText
-                  hasIcon
-                  variant={checkPasswordCondition(
-                    hasLowercaseAndUppercaseLetter,
-                  )}
-                >
-                  Upper & lower case letters
-                </HelperText>
-                <HelperText hasIcon variant={checkPasswordCondition(hasNumber)}>
-                  At least one number
-                </HelperText>
-                <HelperText
-                  hasIcon
-                  variant={checkPasswordCondition(hasSpecialChar)}
-                >
-                  At least special character
-                </HelperText>
-              </div>
+              <PasswordValidation password={password} />
             </div>
           </div>
 
           <div className={field()}>
             <Label>Confirm new password</Label>
-            <Form.Control name="confirmPassword" rules={rules.confirmPassword}>
+            <Form.Control name="confirmPassword" rules={confirmPasswordRules}>
               <Form.PasswordField
                 placeholder="Check the password"
                 autoComplete="one-time-code"
